fix: fix get uncovered block biz

+++ src/main/java/cn/malgo/annotation/request/task/GetUnCoveredBlockRequest.java
@@ -5,5 +5,5 @@ import lombok.Value;
 @Value
 public class GetUnCoveredBlockRequest {
   private int presupposePageSize;
-  private int threshold;
+  private double threshold;
 }
